clean utilis folder
