+++ packages/dappmanager/src/api/auth/sessionAuth.ts
@@ -1,6 +1,6 @@
 import { Request } from "express";
 import { PlainTextFileDb } from "@dappnode/utils";
-import { getRandomAlphanumericToken } from "../../utils/token.js";
+import { getRandomAlphanumericToken } from "./token.js";
 import { wrapHandler } from "../utils.js";
 import { SessionData, SessionsManager } from "../sessions/index.js";
 import { LoginStatusReturn } from "@dappnode/common";
