+++ packages/dappmanager/src/initializeDb.ts
@@ -2,13 +2,15 @@ import * as db from "@dappnode/db";
 import { eventBus } from "@dappnode/eventbus";
 import { generateKeys } from "@dappnode/dyndns";
 import { getDappmanagerImage } from "@dappnode/dockerapi";
-import getServerName from "./utils/getServerName.js";
-import getInternalIp from "./utils/getInternalIp.js";
-import getStaticIp from "./utils/getStaticIp.js";
+import {
+  getInternalIp,
+  getServerName,
+  getStaticIp,
+  ping
+} from "./utils/index.js";
 import { getExternalUpnpIp, isUpnpAvailable } from "@dappnode/upnpc";
 import { writeGlobalEnvsToEnvFile } from "@dappnode/db";
 import { params } from "@dappnode/params";
-import ping from "./utils/ping.js";
 import retry from "async-retry";
 import { IdentityInterface } from "@dappnode/common";
 import { logs } from "@dappnode/logger";
