+++ packages/dappmanager/src/utils/getInternalIp.ts
@@ -2,7 +2,7 @@ import isIp from "is-ip";
 import { shell } from "@dappnode/utils";
 import { getDappmanagerImage } from "@dappnode/dockerapi";
 
-export default async function getInternalIp(): Promise<string> {
+export async function getInternalIp(): Promise<string> {
   try {
     const image = await getDappmanagerImage();
     const output = await shell(
