+++ packages/dappmanager/src/utils/ping.ts
@@ -7,7 +7,7 @@ const max = 3;
  * Checks if a given IP resolves
  * @returns true if resolves
  */
-export default async function ping(ip: string): Promise<boolean> {
+export async function ping(ip: string): Promise<boolean> {
   let i = 0;
   while (i++ < max) {
     // The command ping returns code 0 if some package was successful
