+++ packages/dappmanager/src/utils/validateBackupArray.ts
@@ -11,7 +11,7 @@ import path from "path";
  *   { name: "keystore", path: "/usr/.raiden/secret/keystore" }
  * ]
  */
-export default function validateBackupArray(backup: PackageBackup[]): void {
+export function validateBackupArray(backup: PackageBackup[]): void {
   const pathsObj: { [backupName: string]: boolean } = {};
   for (const { name, path: _path } of backup) {
     if (pathsObj[name])
